mock: Exclude trivial mocks from coverage
Exclude trivial mock functions in common/mock from coverage accounting,
for a somewhat arbitrary value of "trivial." Test coverage is not
inherently important for any mock functions. The mock functions whose
sole purpose is to enable a test build to link are particularly
unimportant. Test authors may expect that the less trivial mocks will be
called during tests, and a lack of coverage might be a useful signal.
Therefore, leave some mocks in the coverage accounting. Reduce the toil
of submitting trivial mock changes to enable tests to link.

BUG=b:433547376
TEST=make runhosttests

Change-Id: I1c0098bcda2b5a3fddf942fd02a037908eb1fa0b