bluey/src/chipset: Cache battery parameters before AP power-on
Once the Application Processor (AP) boots, control of the I2C bus is
relinquished to it, rendering the EC unable to access the battery pack.

To ensure the AP has the data needed for charging decisions, this commit
introduces a mechanism to cache essential battery parameters before the
AP is powered on. This allows both static and dynamic host commands to
fetch the cached values from the EC for the AP to use.

BUG=b:439820798, b:437785464
TEST=zmake build %bluey && confirmed that the AP can access this cached
values using the HOST commands.

Change-Id: I6cc50d57bc102a88e633613d953c6596ca7e8b8a