common: Make host_clear_events_b() a public function
This change modifies host_clear_events_b() from a static function to a
public one. The function is also added to host_command.h to make it
accessible to other modules. This allows for direct manipulation of
the B event mask from other parts of the code.

The B event mask is used by the AP to track the state of its events
(AC status), and this change provides a direct way to clear the events
it has processed.

BUG=b:439819922
TEST=Able to build and boot google/quenbi.

w/o this patch:
 - `hostevent-b` value after `apshutdown` is 0x000000000000018

w/ this patch:
 - `hostevent-b` value after `apshutdown` is 0x000000000000000

Change-Id: I560aae85b83cee550171e9a58f3c9ab42e5e8311