zephyr/bluey: Update AC event status on boot
This commit adds a call to extpower_handle_update() during the
HOOK_CHIPSET_STARTUP hook.

By checking the state of the GPIO_AC_PRESENT pin and calling
extpower_handle_update(), the EC ensures that the AC event status is
correctly initialized and reported to the AP immediately upon chipset
startup. This prevents the AP from having stale or incorrect information
about the power state at the beginning of the boot process.

BUG=b:439819922
TEST=Build and boot google/quenbi.

w/o this patch:
  - AC not present.
  - Power on using power-button.
  - System at charging loop even no AC present.

w/ this patch:
  - AC not present.
  - Power on using power-button.
  - System enter into charging loop and issue shutdown due to no AC
    present.

Change-Id: I5fee1afaa9b201e22f6530828e70af9b11edfce0