[charge_state] Update check for min mW in charge_prevent_power_on
Currently there is an issue where if the user has a charger that
does not meet the minimum platform boot requirements the system
can still boot if the power button is pressed more than 1 time.
This patch removes the check for prevent_power_on before checking
the system power boot requirements. This can lead to constant brown
outs as the system could potentially be unable to stay in a powered
on state without the correct system power.

BUG=none
TEST=verified that ptlrvp_mchp does not boot with 45W charger and
	no battery attached

Change-Id: I27253d6685024b78b46f9cef8b4c9cf801c0a39c