bluey: Configure EC_PMK_RST_OD as push-pull
The EC_PMK_RST_OD signal was originally configured as open-drain (OD).
This configuration was chosen to potentially save a FET inverter on
hardware designs that might utilize an external pull-up resistor.

However, the hardware design requirement was updated to use a pull-down
configuration for this signal. A signal using pull-down must be driven
by a push-pull output, as an open-drain output cannot drive the line
high.

This patch updates the EC_PMK_RST_OD configuration from open-drain to
push-pull to match the hardware requirement.

To reflect this functional change, the signal is also renamed from
EC_PMK_RST_OD to EC_PMK_RST.

This change is backward compatible, for all existing boards/versions

BUG=b:437979589
TEST=zmake build %bluey

Change-Id: I01c1ee3eacd2b760373df12f5df3fa21f4e64264