make: delete the entire build tree when 'make clean' is invoked
This is a bit of a hardhanded change, but it should work fine for the
dev workflow, when CTS_MODULE is not set. Let's delete the root
directory where the board code goes in, when invoking 'make clean',
this has been an assumption all along anyways.

BUG=b:444015384
TEST=running 'make clean' now deletes the entire ./build directory

Change-Id: I36d1d24c112ec3387314da7c4e38cf2e6d2a6332
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6932724
Auto-Submit: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Commit-Queue: Vadim Sukhomlinov <sukhomlinov@google.com>
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@google.com>