Update Microchip WP code to properly protect the flash
The internal Microchip flash was unable to be protected with the existing
code as the internal and externally exposed pins are not the same. Add the
ability to only change the externally exposed pin and keep the internal
pin in alignment with this pin to allow for WP to function properly.

BUG=b:388289782
TEST=zmake ptlrvp_mchp, verify that flashrom is able to protect flash
LOW_COVERAGE_REASON=b:444042320 Microchip will add test code

Change-Id: Ib8bf676763525fbf3de6d4668c8b5099a6c65d5e