tcpci: Enable BIST Test Mode ASAP
In the TCPCI driver, enable BIST Test Mode immediately after receiving
a BIST Test Data message, if conditions are appropriate.
Do not wait for the message to make its way through the PRL and PE.
Decrease the latency of entry into BIST Test Mode, thereby reducing the
opportunity for the TCPC's Rx buffers to overflow during a BIST test.

This generalizes a previous solution that was specific to the RAA489000
TCPC driver.

BUG=b:442310524
TEST=TEST.PD.PHY.ALL.7 Valid SOP*

Change-Id: I286d8b9ed9848976963b4da2bbe25c0598ffe097