pdc_power_mgmt: Don't enable sink paths if dedicated charger attached
PDC power management has logic to ensure multiple sink paths do not get
enabled across PDC chips, but it does not have a way to prevent a sink
path from enabling while a dedicated charge port (such as the RVP barrel
jack) is feeding power into the system.

This should not happen under normal operation, since the sink path
states follow the active charge port, which would be the barrel jack and
not any PDC. However, this mechanism is needed in case a charge port
override is issued. (charge_manager does not have the ability to shut
off the barrel jack load switch)

BUG=b:444320818
TEST=Test on an RVP board

Change-Id: I1938d277b97d55cc3e3b0f4ae2a9f83079d4235e