servo: Introduce SuzyQ alike mode in poweron_conf and CC settings
The original default advertised Type-C current on CC lines,
set to the USB level, is not enough to power some devices
from a dead-battery state.

Therefore we want to introduce a way to support these devices avoiding a risky,
fleet-wide change to the default CC line configuration, which is crucial
for Servo to be properly detected by GSC.

With this change we allow to overwrite default servo CC lines settings
at servo poweron:
    - first inspect the PD SrcCaps of the CHG port - If a fixed PDO
      offering at least 1.5A / 5V is present
    - second check if a special flag is set in nonvolatile memory
    - if both TRUE, we advertise 1.5A on DUT port. This aligns behavior
      with the SusyQ cable, which also advertises 1.5A
    - if ANY FALSE, we stays with safe default USB current level.

This CL also extends current poweron_conf tooling to support this new
flag.

BUG=b:446161577
TEST=Experiment with flag setting/unsetting and TI50 DUT, observe host
     dmesg and ADC measurements/logs in GSC FW console

Change-Id: I78b8b96a07ed8b565d0df521fed45706acb8643c