uldrenite: Enable ISH crash reporting
Intel is providing a new BUP+SHIM for the ISH that does not clear the
memory. In order to accommodate for this we need to artificially shrink
the sram in the overlay. This is because the BUP+SHIM will be loaded
at 0x90000 and if the panic_info struct is written in the normal
location 0xa0000 - sizeof(panic_info) then it will be replaced with the
BUP+SHIM.

Once we write the panic_info to a region that doesn't get cleared during
a reboot, we need to flush the caches.

BUG=b:439591195
TEST=Update the panicinfo host command to cause a crash on the 3rd time
     it executes.
     Flash uldrenite, run `ectool --name cros_ish panicinfo` 3 times.
     The 4th time should shows the panic info from the 3rd

Change-Id: I44cb6443120e7bb6918d7f7ab033beb0c9c52f10