usbc_intr_task: Commands for debugging interrupt storm
The interrupt storm detection feature disables pd ports that are
experiencing an interrupt storm. It is not obvious what the detection
threshold should be for each board. This change adds commands for
debugging and tuning this threshold.

Adds a command to adjust the interrupt storm detection threshold.

Adds a command to simulate an interrupt storm.

BUG=b:436867728
TEST=Deployed and tested on buggzy

Change-Id: Id681392b438107cfc02c94db797a36e781073165