cbi: Add Unified Firmware and Second-source Config (UFSC)
This change implements the next generation of firmware config, as
outlined in go/al-fw-config and go/ec-ufsc. This new UFSC unifies
the legacy FW_CONFIG and SSFC mechanisms into a single, schema-driven
framework. The current ad-hoc, per-board approach is being replaced to
improve consistency and scalability.

This change adds the CBI_TAG_UFSC tag to CBI and its corresponding
struct cbi_ufsc, which is a 5-DWORD value. A new accessor function,
cbi_get_ufsc(), is added to cbi_common.c to retrieve this data.

Unit tests have been added to the drivers.common_cbi suite to validate
the new functionality.

BUG=b:446733213
TEST=./twister -v -T zephyr/test/drivers -s drivers.common_cbi

Change-Id: If00b6b177eabf5d8f933d1d6f08a255c865b7025