cbi-util: Support UFSC field
This change updates the cbi-util tool to support the Unified Firmware
and Second-source Config (UFSC) tag. UFSC was recently introduced as
part of the next-generation firmware configuration to unify legacy
FW_CONFIG and SSFC mechanisms. This cbi-util support is useful for
creating and inspecting CBI images that utilize the new UFSC data
structure during development and manufacturing.

BUG=b:446733213
TEST=make -C util cbi-util

Change-Id: Icd2acafb52b19dc36f35012b6453d2fa4a0beb5c