uldrenite: Add a specific check for 180 SKU
Discovered that the battery cut-off hotkey combination fails on uldrenite.
The root cause is that after an EC reset, the EC incorrectly identifies the device as being in tablet mode and locks the keyboard. This prevents the refresh key press from being detected.
This patch modifies the keyboard locking behavior for the 180 SKU to resolve the issue.

BUG=b:446605593
TEST=cut off success on 180SKU with hot key

Change-Id: I8e9f04af458dd84acf727975eb90e41ac6af36f4