cbi: Add driver for UFSC from device tree
This change introduces the Zephyr driver for the Unified Firmware and
Second-source Config (UFSC). This driver reads a `cbi-ufsc` node from
the device tree to determine the location and size of various bit
fields within the 5-DWORD UFSC block stored in CBI.

This new driver provides a single public API,
`cros_cbi_ufsc_check_match()`, which allows other drivers to check for
the presence of configuration.

Unit tests have been added to the drivers.common_cbi suite to validate
the functionality.

BUG=b:446733213
TEST=./twister -v -T zephyr/test/drivers -s drivers.common_cbi \
-s drivers.common_cbi.system_locked -s drivers.common_cbi.ec_host_cmd

Change-Id: Iadc62ac78aedc3295165e352f0c691f681536b4f