skywalker: support pmic_resetb forwarding
On Skywalker hardware v1.4 and newer, the PMIC_RESETB signal is wired
to the EC. The EC must now forward this signal to EN_PWR_S3 to ensure
correct power sequencing.

This change adds a Kconfig option (`CONFIG_BOARD_SKYWALKER_PMIC_RESETB_IRQ`)
to enable an interrupt handler that mirrors the PMIC_RESETB signal to
the EN_PWR_S3 GPIO. This is enabled by default on boards that have the
`gpio_en_pwr_s3` node defined in their device tree.

BUG=b:373807923
TEST=Build skywalker with following DT change:
	named-gpios {
		gpio_en_pwr_s3: en_pwr_s3 {
			gpios = <&gpioc 7 GPIO_OUTPUT>;
		};
	};

	gpio-interrupts {
		pmic-ec-resetb {
			handler = "pmic_ec_resetb_interrupt";
		};
	};

Change-Id: I92dc15052171b659c98d8aa0be88a6b050bd915f