quenbi: Decouple 5V rail control from the fan driver
The fan driver was incorrectly configured to control `ec_en_pp5000`, the
GPIO that enables the board's +5V power rail. This rail is shared with
the USB-A port.

As a result, commands like `fanduty 0 0` would disable the fan and
inadvertently cut power to the USB-A port, causing it to fail.

This change fixes the issue by removing the `enable-gpio` property from
the fan's device tree node.

Additionally, this patch refactors the rail's startup control. The
previous cros ec shimmed implementation for enabling the 5V rail was
not being triggered. That legacy method is now removed, and the rail
is instead enabled manually during chipset power hooks to guarantee its
availability.

BUG=none
TEST=zmake build %bluey --clobber

w/ this patch:

 ec:> fanduty 0 0
      setting fan 0 duty cycle to 0%
 ec:> fanduty 1 0
      setting fan 1 duty cycle to 0%
 ec:> gpioget ec_en_pp5000
      0  ec_en_pp5000

w/o this patch:
 ec:> fanduty 0 0
      setting fan 0 duty cycle to 0%
 ec:> fanduty 1 0
      setting fan 1 duty cycle to 0%
 ec:> gpioget ec_en_pp5000
      1*  ec_en_pp5000

Change-Id: Ibfd5967a394d702c8b643bee70d3e5d58b50573b