hwtest: Add prj_hw_tests.conf to otp_key test
Include the common `prj_hw_tests.conf` in the `otp_key` test case.

This is necessary to satisfy Kconfig dependencies for
`CONFIG_HW_TEST_OTP_KEY`. Specifically, this option depends on
`CONFIG_HW_TEST`, which is enabled in `prj_hw_tests.conf`.

Without this inclusion, `CONFIG_HW_TEST_OTP_KEY` defaults to 'n' despite
being set to 'y', leading to build warnings and causing the test to hang
as the core test code is not compiled in.

BUG=b:448199791
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t otp_key
     --zephyr

Change-Id: I0857be84b1760a6b1e68e07aa29a3963851252ea