charge_manager: Use mutexes for PDC based devices only
The introduction of mutexes has caused sequencing issues with legacy
TCPM implementations, leading to difficult to debug issues.  The
breaking change always leads to the introduction to the mutex in
charge_manager.  Limit the mutex change for PDC based devices only as
this corrected issues with querying the active charge port.

BUG=b:446729356
TEST=zmake build brox
TEST=zmake build craask

Change-Id: I2e7ca4dd447971e60fe7ae9c007e5b65c3962f61