Zephyr: Enable CONFIG_PLATFORM_EC_SHA256_HW_ZEPHYR default value to be overridden
Add a prompt string to the CONFIG_PLATFORM_EC_SHA256_HW_ZEPHYR Kconfig
symbol. When a symbol is only defined with `bool` and no prompt, the
Zephyr Kconfig system prevents it from being assigned or overridden in
the configuration files (like overlay files or test `.conf` files).

This change allows this config to be disabled (e.g., setting it to `=n`
for the `sha256_unrolled` test), allowing the SHA256 hardware
accelerator to be disabled when needed for test purposes.

BUG=b:448193193
TEST=zmake -D build helipilot --clobber, and make sure the verification
     of the RW image occurs within ~146 msec

Change-Id: If284fc48dbef5858c3fb23c17ca9dedc289d1421