ectool: Add capture type argument to fpframe
The `fpframe` command requires knowledge of the capture type to
correctly interpret the frame data, as different types (e.g., vendor,
pattern0, qual) can have different sensor image metadata, including
frame sizes.

Previously, `ectool` attempted to query the current mode from the EC.
However, the EC firmware clears the transient capture mode immediately
after image acquisition. This makes it unreliable for a subsequent
`fpframe` invocation to determine the capture type used for the stored
frame.

This change enables the user to explicitly specify the expected capture
type as an argument to the `fpframe` command. This ensures the correct
metadata is used when downloading the frame, regardless of the EC's
current mode state.

The following capture type arguments are now supported:
  - vendor
  - pattern0
  - pattern1
  - qual
  - test_reset
  - test_defect_pixel
  - test_abnormal
  - test_noise

If no type is specified, 'vendor' is assumed. The 'raw' argument can
still be used in conjunction with a capture type.

BUG=b:398899644
TEST=emerge-rex ec-utils
TEST=fpframe --> idx = FP_FRAME_INDEX_SIMPLE_IMAGE &&
        capture_type = FP_CAPTURE_VENDOR_FORMAT
TEST=fpframe raw --> idx = FP_FRAME_INDEX_RAW_IMAGE
            capture_type = FP_CAPTURE_VENDOR_FORMAT
TEST=fpframe pattern1 --> idx = FP_FRAME_INDEX_SIMPLE_IMAGE
                 capture_type = FP_CAPTURE_PATTERN1
TEST=fpframe raw test_reset --> idx = FP_FRAME_INDEX_RAW_IMAGE
                       capture_type = FP_CAPTURE_RESET_TEST
TEST=Emerge the `ec-utils` with the current change and cros deploy on a
     device. Capture a `test_reset` image by running `ectool --name=cros_fp
     waitevent FINGERPRINT 500 fpmode capture test_reset. Download the
     image using `ectool --name=cros_fp fpframe > /tmp/test_reset.pnm`.
     Check that the capture and download are successful. Repeat for
     `pattern0`, `pattern1`, and `qual`. For `qual` download the image
     as `raw`.

Change-Id: I550937b037c9c437684f7cdf7bbf875b78d041f4