pdc: Skip sending SET_SINK_PATH command when already set
When battery is cutoff and a low voltage adapter is attached, EC will
brown out when attempting to set the sink path.

PDC will set the supplier for of low voltage adapter, although
charge manager refresh doesn't meet the criteria to select this port
itself. The battery will remain in cutoff mode as the adapter doesn't
provide enough voltage. The PDC will report the sink path is enabled
as coming up from dead battery. Skip sending the sink path when the
desired state is already met.

BUG=None
TEST=cutoff battery, plug in 5v adapter
TEST=./twister -c -T zephyr/test/pdc
TEST=./twister -c -T zephyr/test/charger

Change-Id: If22e0e72634354a1954e16f577fe62cab5fcff24