crisis_rcvry_util.py: Increase retry count
Getting about 10 "CRC mismatch" errors when flashing the EC on
satlab-*-ptl-rack1-bay27. Not sure if this is board specific
or a quirk of satlab hosts. In any case, the original retry
count of 3 is not practical. Let's try 100 since we really want
best effort here and it's low enough that we don't get stuck
trying to flash in an eventual failure case.

TODO: it doesn't look like such a flash error is propagated to an exit
code of flash_ec.

Change-Id: I4b0ee16ea3e9b1d9b2b95858176d9f80d8345084