[TEST]: Riven: Add 4 additional buttons
Add camera, back, overview, and laucher buttons.
For back, overview and laucher buttons are sent the scancode
to inform AP, force to clamshell mode to prevent OS ignored
scancodes in tablet mode.

For camera buttons is support capture image when camera app
is opened in tablet mode, force to tablet mode when the
button is pressed.

This CL let system leave force mode when the lid angle changed
and also disable keyboard scan when force to clamshell mode
for back, overview and launcher buttons at the DUT actually
in tablet mode.

BUG=none
TEST=none

Change-Id: Ib3484f4f37ae74955842769997ade48b0a527e77