fpsensor: Add frame_size to sensor context
This change introduces the `frame_size` member to the `fpsensor_context`
struct. This new field is intended to store the size, in bytes, of the
image data obtained from the most recent fingerprint sensor capture.

The `frame_size` is initialized to 0. This will allow components to
know the buffer size required for or produced by the last capture.

BUG=b:450383638
TEST=CQ

Change-Id: Ia0bfba04c50b148f076958b016fd95c6fa115e38