validate fips section hash at build time.
FIPS section hash is not supposed to change when code changes are
introduced. Let's add a check that the calculated hash matches the
expected value and fail the build right away in case of mismatch.

The expected values are copied from the build tree after a successful
build.

Builds invoked with nonzero CR50_DEV=1 and or CRYPTO_TEST will not
perform the check.

BUG=None
TEST=verified that the error is generated in case one of the submitted
     samples is modified, only when CR50_DEV is not set.

Change-Id: I3ce2e8067f0e2deff774879234519805abea6625