clean up make command output
When running make in Cr50 with V not declared in the environment some
unexpected messages show up in the log:

     RW/ec.RW.hex
   sanitizing util/signer/ec_RW-manifest-TOT.json

This is repeated twice, for each RW image. Another cosmetic issue is
that the log is not aligned properly and some titles are longer than
the most of the rest, which causes uneven log stream.

This patch suppresses the unexpected messages and makes sure that all
titles are printed with the same alignment.

BUG=None
TEST=make buildall succeeds. Make output looks much cleaner.

Change-Id: I15a4ec9d79d1728b2fb656c5c4d845c600ec9d7f