fpsensor: Add num_params to fp_sensor_get_info_v2
Modify the `fp_sensor_get_info_v2` function to accept an in/out
parameter `uint8_t *num_params`. This allows the caller to specify
the maximum number of `fp_image_frame_params` to fetch and the function
to return the actual number of parameters written to the response. The
`fp_command_info_v2` host command handler uses this new parameter,
passing the `num_capture_types` from the request.

BUG=b:445178465
TEST=CQ

Change-Id: If550d4fcc5106f7604e7a07515e50a453e8ea24f