driver/fingerprint: Respect num_params in fp_sensor_get_info_v2
Update the `fp_sensor_get_info_v2` implementation to honor the input
value of `*num_params` when copying image frame parameters. The function
now copies at most the number of `fp_image_frame_params` requested by
the caller.

The number of elements copied is the minimum of the input `*num_params`
and the total number of available parameter sets. The `*num_params`
output value is updated to reflect the number of items actually copied.

BUG=b:445178465
TEST=CQ

Change-Id: Ife12baf6d03d45a42fa2b0f2ad47b3b51cca7761