[WIP] PDC: update PDC suspend flow
As per the UCSI 3.0 specification SET_CCOM with
cc opernation Disabled is available.
If this bit is set, then the connector shall bring it to
Disabled state by removing all terminations from the
CC pins. It shall be done in accordance with Type-C
specification i.e., for example, when the port transitions to
the Disabled state from Attached.SNK, it shall keep all terminations
on the CC pins removed for a minimum of tErrorRecovery.

Change will send SET_CCOM command to PD driver while
trying to suspend the port. This will help disconnect
the devices attached to the port during suspend.

BUG=b:450136517
TEST=

Change-Id: I9733cac654eb1010da72cd44b67ddaf72d2ef152