fpsensor: Use fixed size for FP_INFO v2 response
Modify the EC_CMD_FP_INFO v2 handler to return a response with a
fixed size. The size is calculated based on the maximum possible
number of capture types (FP_MAX_CAPTURE_TYPES), rather than the
actual number present for a given sensor.

This change ensures a predictable response size for the host, which
is expected by host-side libraries such as libec.

The response buffer is now zero-initialized to avoid returning any
uninitialized memory. A check is also added to ensure the host-provided
buffer is large enough, returning EC_RES_OVERFLOW if it's not.

BUG=b:445178465
TEST=ectool --name=cros_fp fpinfo -> succeeds with fpinfo v2 firmware

Change-Id: I3364e2a87eb81ee6f2b15ef48db1fc22fb374bcc