PDC: add pdc console command to set the cc operation mode
Changes implement pdc console command to send SET_CCOM
to the PDC driver. This adds a functionality for soft
diconnect of the attached device to the USBC port
though the device is physically  attached to the port.
Usefull in remote testing, debug of enumeration and PM
related issues.

BUG=b:450136517
TEST= verify "pdc ccom <port> [RP|RD|DRP|DISABLE]"
command from ec console.Sends SET_CCOM command
to the PDC.

Change-Id: If67a76cb52e57e9ebc6adef501b3dc3b867c6dcb