makefile: make BOARD value always available in submakes
When BOARD=xxx is set in the command line or in the environment, the
value of BOARD is propagated into the submakes through the
environment, but when it is set as the default in Makefile, the value
of BOARD is not propagated.

Let explicitly declare it extern to avoid this discrepancy.

BUG=none
TEST=build time fips checksum validation succeeds even when BOARD is
     not set in the make command line.

Change-Id: I7090c0e99db199d28807817e15400f20696e6564
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7043039
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Tested-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Commit-Queue: Vadim Sukhomlinov <sukhomlinov@google.com>
Auto-Submit: Vadim Bendebury <vbendeb@google.com>
Code-Coverage: Vadim Sukhomlinov <sukhomlinov@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>