qcom_exp: Add power-on request for heartbeat wake
This patch introduces a new power-on request and event to allow
board-specific code to trigger a system power-on from a "heartbeat
wake" event.

A new function, `power_on_req_heartbeat()`, is exposed in the
qcom power header. When called, it sets the `POWER_REQ_ON_HEARTBEAT_WAKE`
request and wakes the CHIPSET task.

The chipset task's state machine maps this new request to the
`POWER_ON_BY_HEARTBEAT_WAKE` event.

This event is then handled in `set_pmic_pwron()` similarly to
`POWER_ON_BY_AC_ON`, ensuring that `passthru_ac_on_to_pmic()` is
called to correctly signal the PMIC to power up the system.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:445857032
TEST=On a board that implements this, trigger the heartbeat wake
mechanism and verify the system powers on using the intended flow.

power on reason in the QCLIB logs:
> PM: PON by CBLPWR

Change-Id: I72512d539bc4aa094149b3e7a75b8c85e036287f