system: Add UNEXPECTED_RESET_DETECTION_RW
Add CONFIG_UNEXPECTED_RESET_DETECTION_RW feature.
Detects unexpected restarts in RW image by
storing a flag in bbram upon entering RW and
clears it when leaving. If the flag is set when
entering the image, an unexpected reset has occured.

Only applied to RW image to avoid conflicts between RO
and RW image.

Used to detect when a watchdog occurs in RW.

BUG=b:420744645
TEST=Verify RW panic flag set after watchdog in RW

Change-Id: I42a31bac055067b96a61263fc04880ced83a59b8