common/rollback: Fix alignment issue in crec_flash_physical_write
The crec_flash_physical_write function checks that the offset, size,
and buffer are aligned to CONFIG_FLASH_WRITE_SIZE. When
CONFIG_FLASH_WRITE_SIZE is set to 0x10, the alignment check fails due
to the buffer not being properly aligned.

This commit adds an attribute to ensure the buffer is aligned to
CONFIG_FLASH_WRITE_SIZE to satisfy the alignment requirement.

This fixes the alignment error encountered during rollback operations.

BUG=b:450720190
TEST=zmake -l DEBUG build em32f967 --clobber -->connect FPMCU to CrOS
         -->check /var/log/biod/biod.LATEST
	 "Entropy has been successfully added."

Change-Id: I16b89316e4ae5fb94b184f2d59e019189115876b