power/qcom_exp: Ignore transient AC_PRESENT toggle during switchcap toggle
When the switchcap power supply is enabled or disabled, it
induces a single, spurious toggle on the AC_PRESENT line. This
transient event lasts for approximately 500ms (based on
oscilloscope measurements).

This spurious toggle can be incorrectly interpreted by the EC as a
valid "AC insert" event. This is especially problematic during
shutdown, as it can be mistaken for a power-on signal, causing
the AP (Application Processor) to turn back on unexpectedly.

This patch introduces a 750ms "ignore" window by temporarily
disabling the interrupt to filter out this single transient event.
The 750ms duration provides a safe margin over the measured
~500ms event time.

BUG=b:442784602
TEST=zmake build %bluey
TEST=`power on` and `power off` the AP multiple times to confirm the AC
toogles are ignored and doesn't cause power-on events
TEST= Increase the delay of 750ms to 10 secs and manually plug and
unplug the AC adapter to verify the extpower state changes cleanly.

Change-Id: I6290b6d26add1f66ff39f615d8dac1f8c565baef